// src/app/core/model/rest/tipus-error-500.ts
export interface TipusError500 {
    /**
     * El códi de l'error.
     */
    codi?: string;

    /**
     * La descripció de l'error.
     */
    desc?: string;

    /**
     * Identificador d'incidencia.
     */
    idIncidencia?: string;

    code?: string;

    message?: string;

    /**
     * La propietat relacionada amb l'error.
     */
    propietat?: string;


}